Split all new Additions into their own files
