+++ bytey_byte_buffer/src/byte_buffer_read/byte.rs
@@ -1,0 +1,18 @@
+use crate::{byte_buffer::ByteBuffer, byte_buffer_read::ByteBufferRead, error::Result};
+
+impl ByteBufferRead for bool {
+    #[inline]
+    fn read_from_buffer(buffer: &mut ByteBuffer) -> Result<bool> {
+        Ok(buffer.read::<u8>()? != 0)
+    }
+
+    #[inline]
+    fn read_from_buffer_le(buffer: &mut ByteBuffer) -> Result<bool> {
+        Ok(buffer.read::<u8>()? != 0)
+    }
+
+    #[inline]
+    fn read_from_buffer_be(buffer: &mut ByteBuffer) -> Result<bool> {
+        Ok(buffer.read::<u8>()? != 0)
+    }
+}
