+++ bytey_byte_buffer/src/byte_buffer_write/byte.rs
@@ -1,0 +1,130 @@
+use crate::{byte_buffer::ByteBuffer, byte_buffer_write::ByteBufferWrite, error::Result};
+use std::slice;
+
+impl ByteBufferWrite for u8 {
+    #[inline]
+    fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        unsafe {
+            buffer.write_slice(slice::from_raw_parts(self as *const u8, 1))?;
+        }
+
+        Ok(())
+    }
+
+    #[inline]
+    fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+
+    #[inline]
+    fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+}
+
+impl ByteBufferWrite for &u8 {
+    #[inline]
+    fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        unsafe {
+            buffer.write_slice(slice::from_raw_parts(*self as *const u8, 1))?;
+        }
+
+        Ok(())
+    }
+
+    #[inline]
+    fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+
+    #[inline]
+    fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+}
+
+impl ByteBufferWrite for i8 {
+    #[inline]
+    fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        unsafe {
+            buffer.write_slice(slice::from_raw_parts(self as *const i8 as *const u8, 1))?;
+        }
+
+        Ok(())
+    }
+
+    #[inline]
+    fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+
+    #[inline]
+    fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+}
+
+impl ByteBufferWrite for &i8 {
+    #[inline]
+    fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        unsafe {
+            buffer.write_slice(slice::from_raw_parts(*self as *const i8 as *const u8, 1))?;
+        }
+
+        Ok(())
+    }
+
+    #[inline]
+    fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+
+    #[inline]
+    fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+}
+
+impl ByteBufferWrite for bool {
+    #[inline]
+    fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        let data = *self as u8;
+        unsafe {
+            buffer.write_slice(slice::from_raw_parts(data as *const u8, 1))?;
+        }
+
+        Ok(())
+    }
+
+    #[inline]
+    fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+
+    #[inline]
+    fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+}
+
+impl ByteBufferWrite for &bool {
+    #[inline]
+    fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        let data = **self as u8;
+        unsafe {
+            buffer.write_slice(slice::from_raw_parts(data as *const u8, 1))?;
+        }
+
+        Ok(())
+    }
+
+    #[inline]
+    fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+
+    #[inline]
+    fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
+        self.write_to_buffer(buffer)
+    }
+}
